Split off old tests as regression tests
There are several test cases that cover old features of drf-filters that
have been made obsolete by improvements to django-filter.
